requirements.txt
pyopenssl>=0.11
requests
six
